Split upscale server request helpers
